change direction of conversion

+++ src/rwkvops.py
@@ -269,8 +269,8 @@ class RWKVCudaOps(RWKVPTOps):
         self.initTensor = lambda x: x.to(dtype=self.dtype if len(
             x.shape) == 2 else torch.float32, device='cuda')
         self.postfunc = lambda x: lambda self, y: x(self, y).cpu().float()
-        self.matvec = lambda x, y: x.mv(
-            y.to(self.dtype)).to(dtype=torch.float32)
+        self.matvec = lambda x, y: x.to(dtype=torch.float32).mv(
+            y)
         self.emptyState = torch.zeros(
             4*layers, embed, dtype=torch.float32, device="cuda")+0.01
